requirements.txt
fastapi[standard]
python-multipart
face-recognition
opencv-python
numpy
motor
python-dotenv
pydantic-settings